+++ Project-Mobile/Assets/Scripts/UI/TapObject.cs
@@ -19,7 +19,7 @@ public class TapObject : MonoBehaviour
     }
 
     // Print how much currency is gained by tapping on screen.
-    public void SetValues(long currency, Sprite currencySprite)
+    public void SetValues(double currency, Sprite currencySprite)
     {
         textCurrency.text = "+" + Formatter.FormatValue(currency);
         imageCurrency.sprite = currencySprite;
